+++ lib/mailchimp_api/resources/interest_category.rb
@@ -18,8 +18,8 @@ module MailchimpAPI
     # https://github.com/rails/activeresource/blob/4accda8bc03ceae0ad626f8cec0b4751e89a58ad/lib/active_resource/associations.rb#L151
     #
     # Using a has_many, there is no way to include the `list_id`, so we just create our own interests method.
-    def interests
-      @interests ||= MailchimpAPI::Interest.find(:all, params: { interest_category_id: id }.merge(prefix_options))
+    def interests(params = {})
+      @interests ||= MailchimpAPI::Interest.find(:all, params: { interest_category_id: id }.deep_merge(prefix_options).deep_merge(params))
     end
   end
 end