+++ test/resources/test_member.rb
@@ -48,6 +48,28 @@ describe MailchimpAPI::Member do
     end
   end
 
+  describe 'notes' do
+    it 'calls lists/:list_id/members/:member_id/notes with all IDs populated' do
+      stub_request(:get, SINGLE_MEMBER_URL + '/notes')
+        .to_return status: 200, body: load_fixture(:member)
+
+      member = MailchimpAPI::Member.find 'm1234', params: { list_id: 'list1234' }
+      member.notes
+
+      assert_requested :get, SINGLE_MEMBER_URL + '/notes'
+    end
+
+    it 'uses provided params' do
+      stub_request(:get, SINGLE_MEMBER_URL + '/notes?offset=123')
+        .to_return status: 200, body: load_fixture(:notes)
+
+      member = MailchimpAPI::Member.find 'm1234', params: { list_id: 'list1234' }
+      member.notes offset: 123
+
+      assert_requested :get, SINGLE_MEMBER_URL + '/notes?offset=123'
+    end
+  end
+
   describe 'POST /members/:member_id/actions/delete-permanent' do
     let(:member) { MailchimpAPI::Member.find 'm1234', params: { list_id: 'list1234' } }