+++ CHANGELOG.md
@@ -8,5 +8,6 @@
 - Support of '/lists/<id>/interest-categories'
 - Support of 'lists/<id>/interest-categories/<id>/interests'
 - Support of '/lists/<id>/members'
+- Support of '/lists/<id>/members/<id>/notes'
 - `.count` support
 - Support of '/lists/<id>/merge-fields'